requirements.txt
Flask==2.2.5
Werkzeug==2.2.2
flask-jwt-extended==4.6.0
Flask-OAuthlib==0.9.6
oauthlib==2.1.0
MarkupSafe==2.1.3
Jinja2==3.1.4
itsdangerous==2.1.2
click==8.1.7
requests==2.32.3
requests-oauthlib==1.1.0
PyJWT==2.8.0
PyMySQL==1.1.1
mysql-connector-python==8.4.0
gunicorn
python-dotenv == 1.1.0